Auto set XRT_START_LOCAL_SERVER
